requirements.txt
fastapi
uvicorn
pydantic
transformers
torch
python-dotenv
huggingface_hub
accelerate
accelerate
accelerate
